Rename checksSelection to hostChecksSelection
