--- assets/js/state/selectors/checksSelection.js
@@ -1,4 +1,0 @@
-export const getCheckSelection =
-  () =>
-  ({ checksSelection }) =>
-    checksSelection;
