+++ assets/js/state/selectors/hostChecksSelection.js
@@ -1,0 +1,4 @@
+export const getCheckSelection =
+  () =>
+  ({ hostChecksSelection }) =>
+    hostChecksSelection;
